+++ src/components/ui/MainContainer.component.jsx
@@ -1,4 +1,4 @@
-import { useMemo } from "react";
+import { useMemo, useState } from "react";
 import inventorySampleData from "../../data/inventoryData.json";
 import ToolSection from "../sections/ToolSection.component";
 import QuickStatsBar from "./QuickStatsBar.component";
@@ -7,19 +7,21 @@ import InventorySection from "../sections/InventorySection.component";
 import FilterBarForm from "../forms/FilterBarForm.component";
 
 function MainContainer() {
-  const inventoryItems = useMemo(() => {
-    return inventorySampleData.records.map((item) => ({
-      id: item.id,
-      ...item,
-    }));
+  const inventoryMemo = useMemo(() => {
+    return inventorySampleData.records.map((item) => (item = { ...item }));
   }, []);
+
+  const [inventoryItems, setInventoryItems] = useState(inventoryMemo);
+  const addInventoryItem = (newItem) => {
+    setInventoryItems((prevItems) => [...prevItems, newItem]);
+  };
   return (
     <main>
       <ToolSection id="stats" title="Quick Stats">
         <QuickStatsBar />
       </ToolSection>
       <ToolSection id="add-item" title="Add Item">
-        <AddInventoryItemForm />
+        <AddInventoryItemForm addInventoryItem={addInventoryItem} />
       </ToolSection>
       <InventorySection
         id="fridge"
@@ -47,7 +49,10 @@ function MainContainer() {
         id="shopping-list"
         title="Shopping List"
         shoppingCart
-        items={inventoryItems.filter((item) => item.NeedRestock === "checked" && item.TargetQty > item.QtyOnHand)}
+        items={inventoryItems.filter(
+          (item) =>
+            item.NeedRestock === "checked" && item.TargetQty > item.QtyOnHand,
+        )}
       />
       <ToolSection id="filter" title="Filter & Sort">
         <FilterBarForm />