+++ CHANGELOG.md
@@ -23,6 +23,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.0.4] - 2026-02-27
 
+### Changed
+
+- Refactor MainContainer to manage inventory state and pass addInventoryItem to AddInventoryItemForm
+
 ### Fixed
 
 - Adds a numerical id to each item in inventoryData